refactor : naming 변경
